+++ src/plugins/relay/api/relay-api-protocol.h
@@ -22,6 +22,8 @@
 #ifndef WEECHAT_PLUGIN_RELAY_API_PROTOCOL_H
 #define WEECHAT_PLUGIN_RELAY_API_PROTOCOL_H
 
+#define RELAY_API_ALLOWED_METHODS "GET, POST, PUT, DELETE"
+
 #define RELAY_API_CB(__command) &relay_api_protocol_cb_##__command
 #define RELAY_API_PROTOCOL_CALLBACK(__command)                          \
     enum t_relay_api_protocol_rc                                        \